requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai-whisper>=20231117
rapidfuzz>=3.0.0
openpyxl>=3.1.0
plotly>=5.18.0
numpy>=1.24.0
streamlit-audiorecorder>=0.0.5
soundfile>=0.12.1
pydub>=0.25.1